ytdl_hook.lua: fix boolean comparision
Make sure the `early` variable is boolean. The issue was that if url have
`ytdl://`, find() would return position `1`. This makes `early` also `1`.
And now we try to compare `early == load_fail`, which for `1 == true` is
false, while we expect it to be true. Make sure we compare booleans.

Fixes unnecessary hook processing when using ytdl:// protocol.